+++ scripts/forks/littlelambocoin_install.sh
@@ -4,8 +4,8 @@
 #
 
 LITTLELAMBOCOIN_BRANCH=$1
-# On 2022-07-28
-HASH=60e010097e66fb9a56926f14e890d4d47f6ac465
+# On 2022-08-20
+HASH=8454a7093626a388364ff7425e1b9400ffbbb15a
 
 if [ -z ${LITTLELAMBOCOIN_BRANCH} ]; then
 	echo 'Skipping LittleLamboCoin install as not requested.'
@@ -16,10 +16,10 @@ else
 	git submodule update --init mozilla-ca 
 	git checkout $HASH
 	chmod +x install.sh
-	# 2022-01-30: pip broke due to https://github.com/pypa/pip/issues/10825
-	sed -i 's/upgrade\ pip$/upgrade\ "pip<22.0"/' install.sh
 	# 2022-07-20: Python needs 'packaging==21.3'
 	sed -i 's/packaging==21.0/packaging==21.3/g' setup.py
+	# Log "Added Coins" at info, not debug level.  See: https://github.com/Chia-Network/chia-blockchain/issues/11955
+    sed -e 's/^        self.log.debug($/        self.log.info(/g' littlelambocoin/wallet/wallet_state_manager.py
 	/usr/bin/sh ./install.sh
 
 	if [ ! -d /chia-blockchain/venv ]; then