+++ scripts/pull_3rd_party_libs.sh
@@ -13,9 +13,9 @@ LIST="
 https://cdn.datatables.net/1.12.1/css/dataTables.bootstrap5.css
 https://cdn.datatables.net/1.12.1/js/dataTables.bootstrap5.js
 https://cdn.datatables.net/1.12.1/js/jquery.dataTables.js
-https://cdn.jsdelivr.net/npm/chart.js@3.8.0/dist/chart.min.js
-https://cdn.jsdelivr.net/npm/chartjs-adapter-luxon@1.1.0/dist/chartjs-adapter-luxon.min.js
-https://cdn.jsdelivr.net/npm/chartjs-plugin-datalabels@2.0.0/dist/chartjs-plugin-datalabels.min.js
+https://cdn.jsdelivr.net/npm/chart.js@3.9.1/dist/chart.min.js
+https://cdn.jsdelivr.net/npm/chartjs-adapter-luxon@1.2.0/dist/chartjs-adapter-luxon.min.js
+https://cdn.jsdelivr.net/npm/chartjs-plugin-datalabels@2.1.0/dist/chartjs-plugin-datalabels.min.js
 https://cdn.jsdelivr.net/npm/jquery@3.6.0/dist/jquery.min.js
 https://moment.github.io/luxon/global/luxon.min.js"